abstracted methods out of main. Can now only interact with methods inside of main helper.
